new file ufl_utils.py
